scraper/requirements.txt
requests>=2.31
httpx>=0.27
beautifulsoup4>=4.12
python-dotenv>=1.0
pytest>=8.0
pytest-mock>=3.12
